#25 include missing doguctl verifying docker stage
Re-include doguctl version verification. The docker build stage
`doguctlbinaryverifier` was previously ignored by docker with buildkit
enabled because the stage is not referenced by the final build. This
commit modifies the make target in such a way that the verification will
be called nonetheless.

As https://docs.docker.com/build/building/multi-stage/#differences-between-legacy-builder-and-buildkit
describes:
>
> For example, given the following Dockerfile:
>
> ```dockerfile
> # syntax=docker/dockerfile:1
> FROM ubuntu AS base
> RUN echo "base"
>
> FROM base AS stage1
> RUN echo "stage1"
>
> FROM base AS stage2
> RUN echo "stage2"
> ```
>
> With BuildKit enabled, building the stage2 target in this Dockerfile
> means only base and stage2 are processed. There is no dependency on
> stage1, so it's skipped.